% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{package}
\name{dod-package}
\alias{dod-package}
\title{A Package for downloading publically available data.}
\description{
The \code{dod} package provides tools for downloading Advanced
Microwave Scanning Radiometer (AMSR), buoy, coastline, CTD (BATS,
BBMP, and GTSPP), meteorological, and topographic data.
}
