% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctd_bats.R
\name{dod.ctd.bats}
\alias{dod.ctd.bats}
\title{Download CTD data from the Bermuda Atlantic Time Series program (BROKEN)}
\usage{
dod.ctd.bats()
}
\description{
NOTE: this function has become useless as of 2024-12-17, when the
author noticed that the BATS website no longer provided an interface suitable for
programmatical downloads. Please visit https://bios.asu.edu/bats/bats-data
and use the mouse to find whatever data you seek. You're on your own, my
friend.
}
\seealso{
Other functions that download files: 
\code{\link{dod.amsr}()},
\code{\link{dod.buoy}()},
\code{\link{dod.coastline}()},
\code{\link{dod.ctd}()},
\code{\link{dod.ctd.bbmp}()},
\code{\link{dod.ctd.gtspp}()},
\code{\link{dod.ctd.itp}()},
\code{\link{dod.met}()},
\code{\link{dod.tideGauge}()},
\code{\link{dod.topo}()}
}
\concept{functions that download files}
