% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctd_bbmp.R
\name{dod.ctd.bbmp.index}
\alias{dod.ctd.bbmp.index}
\title{Download an index of BBMP files}
\usage{
dod.ctd.bbmp.index(
  year,
  server = paste0("https://cioosatlantic.ca/erddap/files/",
    "bio_atlantic_zone_monitoring_program_ctd/",
    "Bedford\%20Basin\%20Monitoring\%20Program"),
  debug = 0
)
}
\arguments{
\item{year}{integer or character value indicating the year of interest.
Note that the year 2025 will not work, because the server supplies
information for that year in a pattern that does not match other years. Since
the pattern appears to be temporary, the present code is not set up
to handle it.}

\item{server}{character value indicating the base name of the server.
This may change over time.}

\item{debug}{an integer value indicating the level of debugging. If
this exceeds 0, then some debugging messages will be printed. This
value is passed down to related functions, but with 1 subtracted
for each pass.}
}
\value{
\code{dod.ctd.bbmp.index} returns a data frame with columns
named \code{year}, \code{ID}, and \code{direction}. These elements may be supplied
as parameters of the same name to \code{\link[=dod.ctd.bbmp]{dod.ctd.bbmp()}}.
}
\description{
Download an index of BBMP files available in a specified year. This is done
in analysing the webpage of the data server, and so it will likely fail if
the format of the webpage is changed. Please inform the developers if you
find such a failure, so they can try to make changes.
}
\author{
Dan Kelley
}
