% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctd_bbmp.R
\name{dod.ctd.bbmp}
\alias{dod.ctd.bbmp}
\title{Download CTD data from the Bedford Basin Mooring Project}
\usage{
dod.ctd.bbmp(
  year,
  ID = NULL,
  direction = "DN",
  server = paste0("https://cioosatlantic.ca/erddap/files/",
    "bio_atlantic_zone_monitoring_program_ctd/",
    "Bedford\%20Basin\%20Monitoring\%20Program"),
  file = NULL,
  destdir = ".",
  age = 0,
  quiet = FALSE,
  debug = 0
)
}
\arguments{
\item{year}{a numeric or character value specifying the year of interest.
If this is not provided, it defaults to the current year.}

\item{ID}{a character value indicating the station name. This is
usually in a form like \verb{001_1}, but for some years also have some
profiles with IDs like \verb{001_01}, as well as \verb{001_1}. Using
\code{\link[=dod.ctd.bbmp.index]{dod.ctd.bbmp.index()}} to discover IDs usually works, but
if that fails, explore the webserver defined by the \code{server}
parameter, to check.}

\item{direction}{a character value indicating the direction
of the cast, either \code{"DN"} (the default) or \code{"UP"}.}

\item{server}{character value indicating the base name of the server.
This may change over time.}

\item{file}{character value giving the name to be used for
the downloaded file.  If this is NULL (which is the default) then
the filename is as on the remote data server.}

\item{destdir}{a character value indicating the directory
in which to store downloaded files.}

\item{age}{a numerical value indicating a time interval, in days.
If the file to be downloaded from the server already exists
locally, and was created less than \code{age} days in the past, it will
not be downloaded again.  Setting \code{age=0} forces a download, so
that existing files will always be updated. By contrast, setting
\code{age} to a negative number prevents the updating of files that
already exist locally, regardless of their age.}

\item{quiet}{a logical value, passed to \code{\link[curl:curl_download]{curl::curl_download()}}, which does
the downloading.  If this is TRUE (the default), then the work is done
quietly. Until version version 0.1.12, this parameter was called \code{silent},
but it was renamed to match the corresponding argument in
\code{\link[curl:curl_download]{curl::curl_download()}}, which may be more familiar to users.}

\item{debug}{an integer value indicating the level of debugging. If
this exceeds 0, then some debugging messages will be printed. This
value is passed down to related functions, but with 1 subtracted
for each pass.}
}
\value{
\link{dod.ctd.bbmp} returns a character value naming the file that was
retrieved. See the \sQuote{Examples} section for an example of a typical
workflow.
}
\description{
\link{dod.ctd.bbmp} can retrieve both index files and data files from the Bedford
Basin Mooring Project (BBMP). Since the naming convention of the data files
may alter from year to year, it is important to start by downloading an
index first, using \link{dod.ctd.bbmp.index} first, to determine a file of
interest; see \sQuote{Examples}.
}
\section{Historical Note}{


Sometime in prior to April 2025, the DFO server and the format
of the data files both changed, and this necessitated changes
to \code{\link[=dod.ctd.bbmp]{dod.ctd.bbmp()}}. Also, for clarity of use, the new
function \code{\link[=dod.ctd.bbmp.index]{dod.ctd.bbmp.index()}} was added then.
}

\examples{
# Download and study first BBMP CTD file of year 2024
if (interactive()) { # pkgdown::build_site() cannot handle downloads
    library(dod)
    # Note: cannot specify year=2025 because the URL is differently constructed
    index <- dod.ctd.bbmp.index(year = "2024")
    destdir <- tempdir(check = TRUE) # get temporary storage
    ctdFile <- dod.ctd("BBMP",
        year = index$year[1], ID = index$ID[1],
        direction = "DN", destdir = destdir
    )
    # Use oce to read, summarize and plot the data.
    library(oce)
    ctd <- read.netcdf(ctdFile) |>
        rename() |>
        as.ctd()
    summary(ctd)
    plot(ctd)
    unlink(destdir, recursive = TRUE) # clean up temporary storage
}

}
\seealso{
Other functions that download files: 
\code{\link{dod.amsr}()},
\code{\link{dod.buoy}()},
\code{\link{dod.coastline}()},
\code{\link{dod.ctd}()},
\code{\link{dod.ctd.bats}()},
\code{\link{dod.ctd.gtspp}()},
\code{\link{dod.ctd.itp}()},
\code{\link{dod.met}()},
\code{\link{dod.tideGauge}()},
\code{\link{dod.topo}()}
}
\author{
Dan Kelley
}
\concept{functions that download files}
