% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctd_itp.R
\name{dod.ctd.itp}
\alias{dod.ctd.itp}
\title{Download CTD data from the Ice-Tethered Profiler program}
\usage{
dod.ctd.itp(
  ID,
  info = FALSE,
  file = NULL,
  destdir = ".",
  age = 0,
  quiet = FALSE,
  debug = 0
)
}
\arguments{
\item{ID}{an integer specifying the ID of the profiler that is of interest.}

\item{info}{a logical value.  If \code{info} is FALSE, which is the default, then
\code{\link[=dod.ctd.itp]{dod.ctd.itp()}} downloads the zip-format file that can be expanded into
a directory that holds a data file for each profile made by the
profiler with the given \code{ID}.  The name of the zip file is returned.
(This file must be unzipped to access the data; \code{\link[=unzip]{unzip()}} is a convenient
way to do this.)
On the other hand, if \code{info} is TRUE, then
nothing is downloaded, and \code{\link[=dod.ctd.itp]{dod.ctd.itp()}} returns a
URL holding information about the profiler.}

\item{file}{character value giving the name to be used for the downloaded
file.  If \code{file} is NULL, which is the default, then the filename on the
server is used.}

\item{destdir}{a character value indicating the directory
in which to store downloaded files.}

\item{age}{a numerical value indicating a time interval, in days.
If the file to be downloaded from the server already exists
locally, and was created less than \code{age} days in the past, it will
not be downloaded again.  Setting \code{age=0} forces a download, so
that existing files will always be updated. By contrast, setting
\code{age} to a negative number prevents the updating of files that
already exist locally, regardless of their age.}

\item{quiet}{a logical value, passed to \code{\link[curl:curl_download]{curl::curl_download()}}, which does
the downloading.  If this is TRUE (the default), then the work is done
quietly. Until version version 0.1.12, this parameter was called \code{silent},
but it was renamed to match the corresponding argument in
\code{\link[curl:curl_download]{curl::curl_download()}}, which may be more familiar to users.}

\item{debug}{an integer value indicating the level of debugging. If
this exceeds 0, then some debugging messages will be printed. This
value is passed down to related functions, but with 1 subtracted
for each pass.}
}
\value{
\code{\link[=dod.ctd.itp]{dod.ctd.itp()}} returns the full pathname of the
downloaded zip file, or a URL of an information site,
depending on the value of \code{info}.
}
\description{
\code{\link[=dod.ctd.itp]{dod.ctd.itp()}} downloads CTD data from the Woods Hole Oceanographic
Institutions Ice-Thethered Profile program
\url{https://www2.whoi.edu/site/itp/}.  Note that this
server does not provide a searchable index, so users will need to visit the website
\url{https://www2.whoi.edu/site/itp/data/} to determine IDs of interest.
}
\seealso{
Other functions that download CTD data: 
\code{\link{dod.ctd.bats}()},
\code{\link{dod.ctd.bbmp}()},
\code{\link{dod.ctd.gtspp}()}
}
\author{
Dan Kelley
}
\concept{functions that download CTD data}
