% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctd.R
\name{dod.ctd}
\alias{dod.ctd}
\title{Download CTD data}
\usage{
dod.ctd(program = NULL, ...)
}
\arguments{
\item{program}{character value naming the program (one of listthem).}

\item{...}{extra arguments passed to \code{\link[=dod.ctd.bats]{dod.ctd.bats()}} or one of the
other functions.

If \code{program} is \code{"BATS"}, call \code{\link[=dod.ctd.bats]{dod.ctd.bats()}}.
If \code{program} is \code{"BBMP"}, call \code{\link[=dod.ctd.bbmp]{dod.ctd.bbmp()}}.
If \code{program} is \code{"GTSPP"}, call \code{\link[=dod.ctd.gtspp]{dod.ctd.gtspp()}}.}
}
\description{
This function downloads CTD data from various programs.
}
\details{
\tabular{ll}{
   \strong{Project} \tab \strong{Program} \cr
   Bedford Basin Mooring Project \tab \code{BBMP} \cr
   Bermuda Atlantic Time Series \tab \code{BATS} \cr
   Global Temperature and Salinity Profile Programme \tab \code{GTSPP} \cr
}
}
