% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met.R
\name{dod.met.index}
\alias{dod.met.index}
\title{Get Index of Canadian Meteorological Stations}
\usage{
dod.met.index(
  name,
  max.distance = 0.1,
  url = "https://dd.weather.gc.ca/climate/observations/climate_station_list.csv",
  debug = 0
)
}
\arguments{
\item{name}{character value use in a name search. The search is
done using \code{\link[=agrep]{agrep()}} with the supplied \code{max.distance} value
and with \code{ignore.case} set to TRUE.}

\item{max.distance}{numerical value passed to \code{\link[=agrep]{agrep()}} in
the station-name search.}

\item{url}{the URL of the source file.  The default value was
tested on May 6, 2024, but this agency may change the file
location, or the format, without notice.}

\item{debug}{numerical value indicating the level of debugging. If
this is 0, the work is done silently. For higher values, some
information will be printed as the work is done.}
}
\description{
Look up information on a Canadian meteorological station, by
a fuzzy search on information stored on a gc.ca website. The main
use is to find a station ID, to be used with \code{\link[=dod.met]{dod.met()}}.
}
\details{
The source file is at the web location named by the \code{url}
parameter. (Please contact the author, if the default value fails.
It was known to work in May of 2024, but this agency might change
the file location or format at any time.) The results are a data
frame with information that ought to be easy to interpret.  For use
with \code{\link[=dod.met]{dod.met()}}, the ID field that is of use is named
\code{"Climate.ID"}.  See \dQuote{Examples} for what might be of
interest for hourly data.
}
\examples{
# Get information on the hourly meteorological datasets available
# for Halifax. Use `names(i)` to discover the other column names
# that might be of interest.
library(dod)
i <- dod.met.index("halifax")
i[, c("Station.Name", "Climate.ID", "HLY.First.Year", "HLY.Last.Year")]

}
\author{
Dan Kelley
}
