% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met.R
\name{dod.met.sounding}
\alias{dod.met.sounding}
\title{Download sounding data}
\usage{
dod.met.sounding(
  id = "71603",
  time = Sys.Date() - 1,
  destdir = ".",
  age = 0,
  debug = 0
)
}
\arguments{
\item{id}{character value indicating the station identifier.  The
default is for a station at Yarmouth, Nova Scotia.}

\item{time}{the time for which data are sought, in \code{POSIXct}, \code{POSIXlt}
or \code{Date} format.  The default is the previous day.}

\item{destdir}{a character value indicating the directory
in which to store downloaded files.}

\item{age}{a numerical value indicating a time interval, in days.
If the file to be downloaded from the server already exists
locally, and was created less than \code{age} days in the past, it will
not be downloaded again.  Setting \code{age=0} forces a download, so
that existing files will always be updated. By contrast, setting
\code{age} to a negative number prevents the updating of files that
already exist locally, regardless of their age.}

\item{debug}{an integer value indicating the level of debugging. If
this exceeds 0, then some debugging messages will be printed. This
value is passed down to related functions, but with 1 subtracted
for each pass.}
}
\value{
\code{dod.met.sounding()} returns the name of the downloadeded file.
}
\description{
Download an atmospheric sounding file from the University of Wyoming
Department of Atmospheric science website at
\url{https://weather.uwyo.edu/upperair/sounding.html}. This server has
changed over time, and the present code for \code{dod.met.sounding()} reflects
its file and URL-query structure as inferred on 2025-08-21.
}
\examples{
# Get most yesterday's data at Yarmouth, Nova Scotia.
destdir <- tempdir() # removed at end to pass CRAN checks
file <- dod.met.sounding(destdir = destdir)
data <- read.csv(file)
names(data) # discover names of data items
plot(data$temperature_C, data$pressure_hPa,
    type = "l",
    xlab = "Temperature [C]", ylab = "Pressure [hPa]", ylim = rev(range(data$pressure_hPa))
)
unlink(destdir, recursive = TRUE)

}
\author{
Dan Kelley
}
