% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tide_gauge.R
\name{dod.tideGauge}
\alias{dod.tideGauge}
\title{Download tide-gauge data}
\usage{
dod.tideGauge(
  ID = NULL,
  start = Sys.Date() - 30,
  end = Sys.Date(),
  variable = "water_level",
  file = NULL,
  destdir = ".",
  debug = 0
)
}
\arguments{
\item{ID}{a character (or possibly integer) value specifying the
numeric identifier of the tide gauge.  This is not the name of
the station; for example, the station at Cedar Key in Florida is
queried by using \code{ID=8727520}.}

\item{start, end}{the start and end days of the desired observations.
This may either be a Date object, or character value that can be
converted to one by \code{\link[=as.Date]{as.Date()}}, e.g. "2023-08-30".
If not provided, \code{start} is set to 30 days before the present day,
and \code{end} is set to the present day.}

\item{variable}{a character value indicating the name of the
variable for which data is to be downloaded.  At present, this
must be either \code{"water_level"} or \code{"predictions"}, with the
former being the default.}

\item{file}{a character value indicating the name to be used
for the downloaded data.  If not provided, this is constructed
as e.g. \code{"tide_N_S_E_V.csv"} where \code{N} is the station number,
\code{S} and \code{E} are the start and end dates written in 8-digit format
(i.e. sans the \code{"-"} characters), and \code{V} is the variable
name.}

\item{destdir}{a character value indicating the directory
in which to store downloaded files.}

\item{debug}{an integer value indicating the level of debugging. If
this exceeds 0, then some debugging messages will be printed. This
value is passed down to related functions, but with 1 subtracted
for each pass.}
}
\value{
\code{\link[=dod.tideGauge]{dod.tideGauge()}} returns the full pathname of the
downloaded file.
}
\description{
\code{\link[=dod.tideGauge]{dod.tideGauge()}} downloads tide-gauge data from NOAA, returning
the name of the downloade file.
}
