% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{monthNumberFromName}
\alias{monthNumberFromName}
\title{Convert month name to number, returning NA if no match}
\usage{
monthNumberFromName(name)
}
\arguments{
\item{name}{character value naming a month, with enough
letters to distinguish it from other months.}
}
\value{
integer giving month number, e.g. "Jan" yields 1.
}
\description{
Convert month name to number, returning NA if no match
}
\author{
Dan Kelley
}
