% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/river.R
\name{riverData}
\alias{riverData}
\title{Download and read a river water-level file}
\usage{
riverData(
  id = "01EJ001",
  region = "NS",
  interval = "daily",
  saveFile = FALSE,
  debug = 0
)
}
\arguments{
\item{id}{character value indicating the ID of the desired station. This
may be discovered by using \code{\link[=riverDirectory]{riverDirectory()}} first. This
defaults to \code{"01EJ001"}, for the Sackville River at Bedford.}

\item{region}{character value indicating the province or territory
in which the river gauge is sited. This defaults to \verb{"NS}.}

\item{interval}{character value, either \code{"daily"} or `"hourly",
indicating the time interval desired. (The first seems to yield
data in the current month, and the second seems to yield data
since over the last 1 or 2 days.)}

\item{saveFile}{logical value indicating whether to save the file for later
use. This can be handy if the goal is to archive data, since the
server does not seem to provide archived information.  Please examine
the code of this function to see how to read the file, which is a bit
tricky.}

\item{debug}{integer indicating the level of debugging information
that is printed during processing.  The default, \code{debug=0}, means
to work quietly.}
}
\value{
a data frame containing the data, with columns named \code{"time"} and
\code{"level"}.
}
\description{
Download and read a river water-level file
}
\examples{
library(oce)
library(dod)
dir <- riverDirectory()
data <- riverData(id = dir$id)
oce.plot.ts(data$time, data$level,
    xaxs = "i",
    xlab = "", ylab = "Level [m]", lwd = 2,
    drawTimeRange = FALSE, grid = TRUE
)
mtext(sprintf("\%s (\%.4fN \%.4fE)", dir$name, dir$latitude, dir$longitude),
    line = 0.2, cex = 0.7 * par("cex")
)

}
\author{
Dan Kelley
}
