% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/river.R
\name{riverDirectory}
\alias{riverDirectory}
\title{Acquire a directory of river water-level files}
\usage{
riverDirectory(url, station, debug = 0)
}
\arguments{
\item{url}{character value indicating the URL of a station listing.
If not provided, this defaults to
\verb{https://hpfx.collab.science.gc.ca/today/hydrometric/doc/hydrometric_StationList.csv}
and this is the only URL on which the code is expected to work.}

\item{station}{character value indicating the name of the station for
which information is sought.  This may be a regular expression. If not
provided, \code{"sackville.*bedford"} is used. Set \code{station} to NULL to
get a listing for all stations.}

\item{debug}{integer indicating the level of debugging information
that is printed during processing.  The default, \code{debug=0}, means
to work quietly.}
}
\value{
a data frame holding columns named
\code{"id"}, \code{"name"}, \code{"latitude"}, \code{"region"} and `"tz", with
one row for each selected station.
}
\description{
For examples, see \code{\link[=riverData]{riverData()}}.
}
\author{
Dan Kelley
}
