% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amsr.R
\name{dod.amsr}
\alias{dod.amsr}
\title{Download Advanced Microwave Scanning Radiometer data}
\usage{
dod.amsr(...)
}
\arguments{
\item{...}{arguments passed to
\code{\link[oce:download.amsr]{oce::download.amsr()}}.}
}
\value{
String indicating the full pathname to the downloaded file.
}
\description{
This function downloads topographic data using
\code{\link[oce:download.amsr]{oce::download.amsr()}}.
}
