% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buoy.R
\name{dod.buoy}
\alias{dod.buoy}
\title{Download buoy data}
\usage{
dod.buoy(program, ID = NULL, destdir = ".", age = age, debug = 0)
}
\arguments{
\item{program}{argument specifying the desired oceanographic
program to download buoy data from. Options include \code{MEDS}, etc.}

\item{ID}{For \code{MEDS} this argument is either a location name
or ID number.}

\item{destdir}{a character value indicating the directory
in which to store downloaded files.}

\item{age}{a numerical value indicating a time interval, in days.
If the file to be downloaded from the server already exists locally,
and was created less than \code{age} days in the past, it will not be
downloaded again.  Setting \code{age=0} (which is the default)
forces a download, so that existing files will always be updated.
By contrast, setting \code{age} to a negative number prevents the
updating of files that already exist locally.}

\item{debug}{an integer value indicating the level of debugging. If
this exceeds 0, then some debugging messages will be printed. This
value is passed down to related functions, but with 1 subtracted
for each pass.}
}
\value{
If \code{index} is TRUE, and \code{program} is \code{"BBMP"} or \code{"BATS"},
return a data frame.  Otherwise, return the name of the downloaded file.
}
\description{
This function downloads buoy data from various programs
including...
}
