% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buoy.R
\name{dod.buoy}
\alias{dod.buoy}
\title{Download buoy data}
\usage{
dod.buoy(program, ID = NULL, destdir = ".", age = age, debug = 0)
}
\arguments{
\item{program}{argument specifying the desired oceanographic
program to download buoy data from. If this is not supplied,
\code{dod.buoy()} will print a list of possibilities.}

\item{ID}{a character value indicating the ID of the instrument.}

\item{destdir}{a character value indicating the directory
in which to store downloaded files.}

\item{age}{a numerical value indicating a time interval, in days.
If the file to be downloaded from the server already exists locally,
and was created less than \code{age} days in the past, it will not be
downloaded again.  Setting \code{age=0} (which is the default)
forces a download, so that existing files will always be updated.
By contrast, setting \code{age} to a negative number prevents the
updating of files that already exist locally.}

\item{debug}{an integer value indicating the level of debugging. If
this exceeds 0, then some debugging messages will be printed. This
value is passed down to related functions, but with 1 subtracted
for each pass.}
}
\value{
If \code{index} is TRUE, and \code{program} is \code{"BBMP"} or \code{"BATS"},
return a data frame.  Otherwise, return the name of the downloaded file.
}
\description{
This function downloads buoy data from various programs
including...
}
\examples{
# Significant wave height measured by a buoy in Halifax Harbour
library(dod)
tmpdir <- tempfile()
print(tmpdir)
dir.create(tmpdir)
file <- dod.buoy("smartatlantic", "h1", destdir = tmpdir)
col.names <- strsplit(readLines(file, 1), ",")[[1]]
d <- read.csv(file, skip = 2, col.names = col.names)
d$t <- as.POSIXct(d$time, tz = "UTC", format = "\%Y-\%m-\%dT\%H:\%M:\%SZ")
plot(d$t, d$wave_ht_sig, type = "l",
    xlab = "", ylab = "Sig. Wave Ht. [m]"
)
unlink(tmpdir, recursive = TRUE)

}
\author{
Dan Kelley
}
