% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coastline.R
\name{dod.coastline}
\alias{dod.coastline}
\title{Download coastline data}
\usage{
dod.coastline(...)
}
\arguments{
\item{...}{arguments passed to
\code{\link[oce:download.coastline]{oce::download.coastline()}}.}
}
\value{
String indicating the full pathname to the downloaded file.
}
\description{
This function downloads topographic data using
\code{\link[oce:download.coastline]{oce::download.coastline()}}.
}
