% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctd.bats.R
\name{dod.ctd.bats}
\alias{dod.ctd.bats}
\title{Download CTD data from the Bermuda Atlantic Time Series (BATS)}
\usage{
dod.ctd.bats(
  year,
  ID = NULL,
  index = FALSE,
  file = NULL,
  destdir = ".",
  debug = 0,
  read = FALSE
)
}
\arguments{
\item{year}{a character value specifying the year of interest.}

\item{ID}{a character value specifying the file of interest.}

\item{index}{a boolean value indicating whether the index
should be downloaded.}

\item{file}{character value giving the name to be used for
the downloaded file.}

\item{destdir}{a character value indicating the directory
in which to store downloaded files.}

\item{debug}{an integer value indicating the level of debugging. If
this exceeds 0, then some debugging messages will be printed.}

\item{read}{a boolean indicated weather or not the downloaded
file should be downloaded}
}
\value{
If \code{read} is FALSE, a downloaded file is returned. If \code{read}
is TRUE a data frame is returned.
}
\description{
This function downloads CTD data from BATS.
}
\examples{
\dontrun{
library(dod)
test <- dod.ctd.bats(index=FALSE, ID=10001, read=TRUE)
head(test)
}
}
