% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctd.bbmp.R
\name{dod.ctd.bbmp}
\alias{dod.ctd.bbmp}
\title{Download CTD data from the Bedford Basin Mooring Project}
\usage{
dod.ctd.bbmp(
  year,
  ID = NULL,
  index = FALSE,
  file = NULL,
  destdir = ".",
  debug = 0
)
}
\arguments{
\item{year}{a character value specifying the year of interest.}

\item{ID}{a character value specifying the file of interest
that is copied from the index.}

\item{index}{a boolean value indicating whether the index
should be downloaded.}

\item{file}{character value giving the name to be used for
the downloaded file.}

\item{destdir}{a character value indicating the directory
in which to store downloaded files.}

\item{debug}{an integer value indicating the level of debugging. If
this exceeds 0, then some debugging messages will be printed.}
}
\value{
\link{dod.ctd.bbmp} returns a character value naming the file that
was retrieved. This may either be an index file or a data file;
see \sQuote{Examples} for a typical workflow in which an index is
retrieved and then used as a guide to the name of a data file of
interest.
}
\description{
\link{dod.ctd.bbmp} can retrieve both index files and data files
from the Bedford Basin Mooring Project (BBMP). Since the naming
convention of the data files may alter from year to year, it is
important to start by downloading an index first, to ascertain
the name of a file of interest; see \sQuote{Examples}.
}
\examples{
\dontrun{
# Download the first file of year 2022.
library(dod)
# Download the index
indexFile <- dod.ctd("BBMP", year=2022, index=TRUE, file="index_bbmp_2022")
# Read the index (altering this call if the file format changes)
index <- read.csv(indexFile, header=FALSE, col.names=c("file","time"), skip=3)
# Download the first file in the index
ctdFile <- dod.ctd("BBMP", year=2022, ID=index$file[1], file="bbmp_2022_001")
# Use oce to read, summarize and plot the data.
library(oce)
ctd <- read.ctd(ctdFile)
summary(ctd)
plot(ctd)
}

}
