% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctd.bbmp.R
\name{dod.ctd.bbmp}
\alias{dod.ctd.bbmp}
\title{Download CTD data from the Bedford Basin Mooring Project (BBMP)}
\usage{
dod.ctd.bbmp(
  year,
  ID = NULL,
  index = FALSE,
  file = NULL,
  destdir = ".",
  debug = 0
)
}
\arguments{
\item{year}{a character value specifying the year of interest.}

\item{ID}{a character value specifying the file of interest
that is copied from the index.}

\item{index}{a boolean value indicating whether the index
should be downloaded.}

\item{file}{character value giving the name to be used for
the downloaded file.}

\item{destdir}{a character value indicating the directory
in which to store downloaded files.}

\item{debug}{an integer value indicating the level of debugging. If
this exceeds 0, then some debugging messages will be printed.}
}
\value{
If \code{index} is TRUE, return a data frame. Otherwise,
return the name of the downloaded file.
}
\description{
This function downloads CTD data from BBMP.
}
\examples{
\dontrun{
# Download the first file of year 2022.
library(dod)
library(oce)
destdir <- "~/data/ctd"
i <- dod.ctd("BBMP", year=2022, index=TRUE, file="index")
f <- dod.ctd("BBMP", year=2022, ID=i$file[1], destdir=destdir, file="bbmp")
ctd <- read.ctd(f)
plot(ctd)
}

}
