% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctd.gtspp.R
\name{dod.ctd.gtspp}
\alias{dod.ctd.gtspp}
\title{Download CTD data from the Bermuda Atlantic Time Series (BATS)}
\usage{
dod.ctd.gtspp(ID = NULL, file = NULL, destdir = ".", debug = 0)
}
\arguments{
\item{ID}{a character value consisting of 8 digits where
the first two are an ocean basin initial (at, pa or in), the
next four are the year of interest and the last two are the
month of interest.}

\item{file}{character value giving the name to be used for
the downloaded file. This does include the extension.}

\item{destdir}{a character value indicating the directory
in which to store downloaded files.}

\item{debug}{an integer value indicating the level of debugging. If
this exceeds 0, then some debugging messages will be printed.}
}
\value{
Return a data frame.
}
\description{
This function downloads CTD data from BATS.
}
