% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctd.gtspp.R
\name{dod.ctd.gtspp}
\alias{dod.ctd.gtspp}
\title{Download CTD data from the Bermuda Atlantic Time Series (BATS)}
\usage{
dod.ctd.gtspp(
  ID = NULL,
  file = NULL,
  destdir = ".",
  debug = 0,
  index = TRUE,
  read = FALSE,
  nc = NULL
)
}
\arguments{
\item{ID}{a character value consisting of 8 digits where
the first two are an ocean basin initial (at, pa or in), the
next four are the year of interest and the last two are the
month of interest.}

\item{file}{character value giving the name to be used for
the downloaded file.}

\item{destdir}{a character value indicating the directory
in which to store downloaded files.}

\item{debug}{an integer value indicating the level of debugging. If
this exceeds 0, then some debugging messages will be printed.}

\item{index}{a boolean value indicating whether the index
should be downloaded.}

\item{read}{a boolean indicated weather or not the downloaded
file should be downloaded}

\item{nc}{the name of a netcdf file to download, likely from
the index}
}
\value{
\enumerate{
\item If \code{index} = TRUE and \code{read}=FALSE a gz file is returned.
\item If \code{index}=TRUE and \code{read} = TRUE, a data frame with a list of
NetCDF files names is returned. 3) If \code{index}=FALSE and
\code{read} = FALSE, a netcdf file of the specified index is returned.
\item If \code{index} = FALSE and \code{read} = TRUE, a read netcdf file is
turned.
}
}
\description{
This function downloads CTD data from BATS.
}
