% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctd.R
\name{dod.ctd}
\alias{dod.ctd}
\title{Download CTD data}
\usage{
dod.ctd(program, year, ID = NULL, index = FALSE, destdir = ".", debug = 0)
}
\arguments{
\item{program}{a character value specifying the oceanographic
program from which the data derive (see \sQuote{Details}).}

\item{year}{a character value specifying the year of interest.}

\item{ID}{a character value specifying the file of interest
(see \sQuote{Details}).}

\item{index}{a boolean value indicating whether the index
should be downloaded.}

\item{destdir}{a character value indicating the directory
in which to store downloaded files.}

\item{debug}{an integer value indicating the level of debugging. If
this exceeds 0, then some debugging messages will be printed.}
}
\value{
If \code{index} is TRUE, and \code{program} is \code{"BBMP"} or \code{"BATS"},
return a data frame.  Otherwise, return the name of the downloaded file.
}
\description{
This function downloads CTD data from various programs.
}
\details{
\tabular{llll}{
   \strong{Project} \tab \strong{Program} \tab \strong{Index} \tab \strong{ID} \cr
   Bedford Basin Mooring Project \tab \code{BBMP} \tab Yes \tab From index \cr
   Bermuda Atlantic Time Series \tab \code{BATS} \tab Yes \tab Cruise ID \cr
}
}
\examples{
\dontrun{
# Download the first file of year 2022.
destdir <- "~/data/ctd"
i <- dod.ctd("BBMP", year=2022, index=TRUE)
f <- dod.ctd("BBMP", year=2022, ID=i$file[1], destdir=destdir)
library(oce)
ctd <- read.ctd(f)
plot(ctd)
}

}
