% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctd.R
\name{dod.ctd}
\alias{dod.ctd}
\title{Download ctd data}
\usage{
dod.ctd(program, year, ID = NULL, index = FALSE, destdir = ".", debug = 0)
}
\arguments{
\item{program}{argument specifying the desired oceanographic
program to download ctd data from. Options include BBMP, etc.}

\item{year}{argument specifying the year of interest.}

\item{ID}{for BBMP, this is the desired file from the index.
For BATS, is the 5 digit cruise ID.}

\item{index}{a boolean value indicating whether the index
should be downloaded.}

\item{destdir}{character value indicating the directory
in which to store downloaded files.}

\item{debug}{integer value indicating level of debugging.
If this is less than 1, no debugging is done. Otherwise,
some functions will print debugging information.}
}
\value{
If \code{index} is TRUE, and \code{program} is \code{"BBMP"} or \code{"BATS"},
return a data frame.  Otherwise, return the name of the downloaded file.
}
\description{
This function downloads ctd data from various programs
including the Bedford Basin Mooring Program (BBMP), etc.
}
\examples{
\dontrun{
# Download the first file of year 2022.
destdir <- "~/data/ctd"
i <- dod.ctd("BBMP", year=2022, index=TRUE)
f <- dod.ctd("BBMP", year=2022, ID=i$file[1], destdir=destdir)
library(oce)
ctd <- read.ctd(f)
plot(ctd)
}

}
