% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctd.R
\name{dod.ctd}
\alias{dod.ctd}
\title{Download ctd data}
\usage{
dod.ctd(program, year, ID = NULL, index = FALSE, destdir = ".", debug = 0)
}
\arguments{
\item{program}{argument specifying the desired oceanographic
program to download ctd data from. Options include BBMP, etc.}

\item{year}{argument specifying the year of interest.}

\item{ID}{for BBMP, this is the desired file from the index.
For BATS, is the 5 digit cruise ID.}

\item{index}{a boolean value indicating whether the index
should be downloaded.}

\item{debug}{integer value indicating level of debugging.
If this is less than 1, no debugging is done. Otherwise,
some functions will print debugging information.}
}
\value{
If program = "BBMP", an \code{oce} object is returned.
If program = "BATS", a data frame is returned.
}
\description{
This function downloads ctd data from various programs
including the Bedford Basin Mooring Program (BBMP), etc.
}
