% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctd.R
\name{dod.ctd}
\alias{dod.ctd}
\title{Download ctd data}
\usage{
dod.ctd(program, year, item, debug = 0)
}
\arguments{
\item{program}{argument specifying the desired oceanographic
program to download ctd data from. Options include BBMP, etc.}

\item{year}{argument specifying the year of interest.}

\item{item}{the type of data the user wishes to download.
For example, \code{index} or \code{data}.

@param debug integer value indicating level of debugging.
If this is less than 1, no debugging is done. Otherwise,
some functions will print debugging information.}
}
\description{
This function downloads ctd data from various programs
including the Bedford Basin Mooring Program (BBMP), etc.
}
