% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{dod.download}
\alias{dod.download}
\title{Download with diagnostics}
\usage{
dod.download(url, file, destdir)
}
\arguments{
\item{url}{link generated by the code?}

\item{file}{object containing data?}

\item{destdir}{character value indicating the directory
in which to store downloaded files.}
}
\description{
Download with diagnostics
}
