% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{dod.download}
\alias{dod.download}
\title{Download a file with error checking}
\usage{
dod.download(url = NULL, file = NULL, destdir = ".", silent = TRUE, debug = 0)
}
\arguments{
\item{url}{character value giving the web address of a file to be
downloaded. This has different values for different file types.}

\item{file}{character value giving the name to be used for
the downloaded file.}

\item{destdir}{character value indicating the directory
in which to store the downloaded file.}

\item{silent}{logical value passed to \code{\link[=download.file]{download.file()}}, which
does the downloading.  The default, TRUE, indicates not to show
a progress indicator.}

\item{debug}{integer giving the level of debugging. If
this exceeds zero, then some information is printed about
the processing.}
}
\value{
\code{dod.download} returns a character value holding the full
name of the file, including the path to \code{destdir}.
}
\description{
\code{dod.download()} is mainly intended for use by other functions
in the \code{dod} package.
}
