% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{dod.download}
\alias{dod.download}
\title{Download with diagnostics}
\usage{
dod.download(url, file, destdir)
}
\arguments{
\item{url}{hello}

\item{file}{hello}

\item{destdir}{hello}
}
\description{
Download with diagnostics
}
