% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met.R
\name{dod.met}
\alias{dod.met}
\title{Download meteorological data}
\usage{
dod.met(...)
}
\arguments{
\item{...}{arguments passed to
\code{\link[oce:download.met]{oce::download.met()}}.}
}
\value{
String indicating the full pathname to the downloaded file.
}
\description{
This function downloads topographic data using
\code{\link[oce:download.met]{oce::download.met()}}.
}
