% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met.R
\name{dod.met}
\alias{dod.met}
\title{Download meteorological data}
\usage{
dod.met(...)
}
\arguments{
\item{...}{additional arguments passed to
\link[oce:download.met]{oce::download.met}.}
}
\description{
This function downloads topographic data using
\link[oce:download.met]{oce::download.met}.
}
