% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met.R
\name{dod.met}
\alias{dod.met}
\title{Download meteorological data}
\usage{
dod.met(...)
}
\arguments{
\item{...}{arguments passed to \code{\link[oce:download.met]{oce::download.met()}}.}
}
\value{
\link{dod.met} returns a character value holding the full pathname
of the downloaded file.
}
\description{
Get meteorological data from an Environment Canada website, using
\code{\link[oce:download.met]{oce::download.met()}}.
}
