% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dod.R
\name{dod}
\alias{dod}
\title{General function for downloading data}
\usage{
dod(type, ...)
}
\arguments{
\item{type}{a character value indicating the type of data eg. \code{"topo"}}
}
\description{
For example if \code{type} is \code{"topo"}, \code{\link[oce:download.topo]{oce::download.topo()}} is called.
}
\author{
Annie Howard, Jaimie Harbin
}
