% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{package}
\name{dod}
\alias{dod-package}
\alias{dod}
\title{A package for downloading publically available data.}
\description{
The \code{dod} package provides tools for downloading Advanced
Microwave Scanning Radiometer (AMSR), buoy, coastline, CTD (BATS,
BBMP, and GTSPP), meteorological, tidal, and topographic data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://dankelley.github.io/dod/}
  \item Report bugs at \url{https://github.com/dankelley/dod/issues}
}

}
\author{
\strong{Maintainer}: Dan Kelley \email{Dan.Kelley@Dal.Ca} (\href{https://orcid.org/0000-0001-7808-5911}{ORCID})

Authors:
\itemize{
  \item Annie Howard (\href{https://orcid.org/0000-????-????-????}{ORCID})
  \item Jaimie Harbin \email{Jaimie.Harbin@dfo-mpo.gc.ca} (\href{https://orcid.org/0000-0003-3774-3732}{ORCID})
}

}
\keyword{internal}
