% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/river.R
\name{dod.river.usgs}
\alias{dod.river.usgs}
\title{Download American river-gauge data from USGS}
\usage{
dod.river.usgs(id = "03242350", start = NULL, end = NULL)
}
\arguments{
\item{id}{character value giving the numeric code for the gauge. This
defaults to a river in Ohio.}

\item{start}{an indication of the start time of the requested data window.
If provided, this may be in a (non-ambiguous) character form or, better,
in a POSIXt object. The "UTC" timezone is assumed.  If not provided,
\code{start} defaults to 1 week before the present time.}

\item{end}{as for \code{start}, but instead for the end date.}
}
\description{
Download American river-gauge data from USGS
}
\examples{
library(dod)
file <- dod.river.usgs(start = "2025-08-01", end = "2025-08-08")
lines <- readLines(file)
skip <- 1 + grep("agency_cd", lines)
data <- read.delim(text = lines, skip = skip, sep = "\t", header = FALSE)
# Note that the data at this site, if downloaded during daylight-savings
# time, are offset from UTC by 4 hours. I'm not sure how to make the server
# return in UTC.
time <- as.POSIXct(data$V3, tz = "UTC") + 4 * 3600
# Convert from feet to metres
height <- 0.3048 * data$V5
plot(time, height, type = "l")
mtext(paste("Station", data$V2[1]))
file.remove(file) # needed for tests on CRAN

}
\author{
Dan Kelley
}
