% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tide_gauge.R
\name{dod.tideGauge}
\alias{dod.tideGauge}
\title{Download tide-gauge data}
\usage{
dod.tideGauge(
  ID = NULL,
  variable = "water_level",
  agency = "CHS",
  start = NULL,
  end = NULL,
  resolution = NULL,
  file = NULL,
  destdir = ".",
  age = 0,
  quiet = FALSE,
  debug = 0
)
}
\arguments{
\item{ID}{a character (or possibly integer) value specifying the
numeric identifier of the tide gauge. For Canadian data, this
is either the station number or the station name (e.g. number 491
corresponds to the "Bedford Institute" station).  For American
data, it is a numerical code.}

\item{variable}{a character value indicating the name of the variable to be
downloaded.  There are three types.  (1) \code{"water_level"} (the default)
yields observed water level. This is called \code{"wlo"} on the CHS server and
\code{"water_level"} on the NOAA server (on the latter, see also \code{resolution}).
(2) \code{"predictions"} yields tidal predictions. These are called \code{"wlp"} by
CHS and \code{"predictions"} by NOAA. (3) \code{"metadata"} yields a list containing
information about the station, such as its code number, its official name,
its datum, etc.}

\item{agency}{a character value indicating the agency from which data will
be sought. Use \code{"CHS"} for Canadian tide gauges (the default), or \code{"NOAA"}
for American tide gauges.}

\item{start, end}{POSIXt times or character values that can be converted to
such times (via \code{\link[=as.POSIXct]{as.POSIXct()}} with \code{tz="UTC"}) that indicate the time
interval for the requested data.  If \code{end} is not specified, then the
present time is used.  If \code{start} is not provided, then it is set to the
present time minus 7 days.  If \code{start} is a numeric value, then it is
interpreted as the number of days to go back in time from the \code{end} time.}

\item{resolution}{character value indicating the time resolution. This is
handled differently for the two choices of \code{agency}. (1) For \code{agency="CHS"},
the choices are \code{"ONE_MINUTE"}, \code{"THREE_MINUTES"} (the default),
\code{"FIVE_MINUTES"}, \code{"FIFTEEN_MINUTES"}, and \code{"SIXTY_MINUTES"}.  (2) For
\code{agency="NOAA"}, the choices are \code{"ONE_MINUTE"}, \code{"SIX_MINUTES"} and
\code{"SIXTY_MINUTES"} (the default).}

\item{file}{a character value indicating the name to be used for the
downloaded data.  If not provided, this is constructed as e.g.
\code{"tide_A_N_S_E_R_V.csv"} where \code{A} is the value of the agency argument, \code{N}
is the station ID number, \code{S} and \code{E} are the start and end dates written in
8-digit format (i.e. sans the \code{"-"} characters), \code{R} is the resolution in
minutes, and \code{V} is the variable name.}

\item{destdir}{a character value indicating the directory
in which to store downloaded files.}

\item{age}{a numerical value indicating a time interval, in days.
If the file to be downloaded from the server already exists
locally, and was created less than \code{age} days in the past, it will
not be downloaded again.  Setting \code{age=0} forces a download, so
that existing files will always be updated. By contrast, setting
\code{age} to a negative number prevents the updating of files that
already exist locally, regardless of their age.}

\item{quiet}{a logical value, passed to \code{\link[curl:curl_download]{curl::curl_download()}}, which does
the downloading.  If this is TRUE (the default), then the work is done
quietly. Until version version 0.1.12, this parameter was called \code{silent},
but it was renamed to match the corresponding argument in
\code{\link[curl:curl_download]{curl::curl_download()}}, which may be more familiar to users.}

\item{debug}{an integer value indicating the level of debugging. If
this exceeds 0, then some debugging messages will be printed. This
value is passed down to related functions, but with 1 subtracted
for each pass.}
}
\value{
\code{\link[=dod.tideGauge]{dod.tideGauge()}} returns a list, if \code{value} is \code{"metadata"}
and \code{agency} is \code{"CHS"}; otherwise it returns a file name (with
full path included).  For the CHS case this is a constructed filename,
since the CHS server provides data, not files.  For the
NOAA case, it is a downloaded file.
}
\description{
\code{\link[=dod.tideGauge]{dod.tideGauge()}} downloads tide-gauge data from either
Canadian (CHS) or American (NOAA) tide gauges.
}
\details{
Downloads are done from for either the Canadian Hydrographic Service (CHS)
or from the American National Oceanographic and Atmospheric Agency (NOAA),
respectively (see References 1 and 2). The resultant data are saved in
either constructed filenames, or filenames provided by the user; in either
case, if a recent file already exists with the indicated name, then no data
are downloaded.

For NOAA files, water level and predictions are provided on the same time
sequence, but for CHS files, this is not the case, e.g. predictions (in the
author's tests) are on 15 minute intervals, starting at an hour marker, but
observations may be at a variety of times, depending \code{start} and \code{end}.
Therefore, numerical comparision with CHS data will require interpolation
(see \sQuote{Examples}).
}
\examples{
# Download and plot tide-guage data for Halifax Harbour
library(dod)
library(oce)
destdir <- tempfile() # removed at the end, since CRAN requires that
dir.create(destdir)
message(destdir)
ofile <- dod.tideGauge(491, destdir = destdir)
pfile <- dod.tideGauge(491, "predictions", destdir = destdir)
O <- read.csv(ofile)
O$time <- as.POSIXct(O$Date.Time, "\%Y-\%m-\%d \%H:\%M:\%S", tz = "UTC")
P <- read.csv(pfile)
P$time <- as.POSIXct(P$Date.Time, "\%Y-\%m-\%d \%H:\%M:\%S", tz = "UTC")
# Top panel: observation (black) and prediction (gray)
par(mfrow = c(2, 1))
oce.plot.ts(O$time, O$Water.Level, ylab = "Water Level [m]", xaxs = "i")
lines(P$time, P$Predictions, col = "gray", type = "l")
# Bottom panel: misfit. Note the interpolation to observation time.
misfit <- O$Water.Level - approx(P$time, P$Predictions, O$time)$y
oce.plot.ts(O$time, misfit, ylab = "Deviation [m]", xaxs = "i")
unlink(destdir, recursive = TRUE)

}
\references{
\enumerate{
\item \url{https://api.iwls-sine.azure.cloud-nuage.dfo-mpo.gc.ca/swagger-ui/index.html}
\item \url{https://api.tidesandcurrents.noaa.gov/api/prod/}
}
}
\seealso{
Other functions that download files: 
\code{\link{dod.amsr}()},
\code{\link{dod.buoy}()},
\code{\link{dod.coastline}()},
\code{\link{dod.ctd}()},
\code{\link{dod.ctd.bats}()},
\code{\link{dod.ctd.bbmp}()},
\code{\link{dod.ctd.gtspp}()},
\code{\link{dod.ctd.itp}()},
\code{\link{dod.met}()},
\code{\link{dod.topo}()}
}
\author{
Dan Kelley
}
\concept{functions that download files}
