% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tide_gauge.R
\name{dod.tideGauge}
\alias{dod.tideGauge}
\title{Download tide-gauge data}
\usage{
dod.tideGauge(
  ID = NULL,
  variable = "height",
  agency = "CHS",
  start = NULL,
  end = NULL,
  resolution = NULL,
  file = NULL,
  destdir = ".",
  age = 0,
  debug = 0
)
}
\arguments{
\item{ID}{a character (or possibly integer) value specifying the
numeric identifier of the tide gauge. For Canadian data, this
is either the station number or the station name (e.g. number 491
corresponds to the "Bedford Institute" station).  For American
data, it is a numerical code.}

\item{variable}{a character value indicating the name of the variable to
be downloaded.  This defaults to \code{"height"} for observed height (called \code{"wlo"}
on the CHS server and \code{"water_level"} on the NOAA server).  The other
choice is \verb{"heightPredicted" (called }"wlp"\verb{by CHS and}"predictions"\verb{by NOAA). In either case,}time\code{and}QC` are also stored alongside
the variable.}

\item{agency}{a character value indicating the agency from which data will be
sought. Use \verb{"CHS" for Canadian tide gauges (the default), or }"NOAA"` for
American tide gauges.}

\item{start, end}{POSIXt times or character values that can be converted
to such times (via \code{\link[=as.POSIXct]{as.POSIXct()}} with \code{tz="UTC"}) that indicate
the time interval for the requested data.
If either \code{start} or \code{end} is not supplied, then a default
of the past week is used.}

\item{resolution}{character value indicating the resolution, only used
if \code{agency} is \code{"CHS"}.  The choices are \verb{"ONE_MINUTE", }"THREE_MINUTES"\verb{(the default),}"FIVE_MINUTES"\verb{, }"FIFTEEN_MINUTES"\verb{, and }"SIXTY_MINUTES"\verb{.  The server website indicates that }"Not all resolution are available for every station and every time series."`
It also indicates that
"The time frame available increases with the resolution, it is equal
to a week (7 days) multiplied by the resolution, but the maximum
available is 31 days. (a month)".  This second statement perhaps
means that it is possible to download at most 1 week of data for
1-minute resolution.}

\item{file}{a character value indicating the name to be used for the
downloaded data.  If not provided, this is constructed as e.g.
\code{"tide_A_N_S_E_R_V.csv"} where \code{A} is the value of the agency argument, \code{N} is
the station ID number, \code{S} and \code{E} are the start and end dates written in
8-digit format (i.e. sans the \code{"-"} characters), \code{R} is the resolution in
minutes, and \code{V} is the variable name.}

\item{destdir}{a character value indicating the directory
in which to store downloaded files.}

\item{age}{a numerical value indicating a time interval, in days.
If the file to be downloaded from the server already exists locally,
and was created less than \code{age} days in the past, it will not be
downloaded again.  Setting \code{age=0} (which is the default)
forces a download, so that existing files will always be updated.
By contrast, setting \code{age} to a negative number prevents the
updating of files that already exist locally.}

\item{debug}{an integer value indicating the level of debugging. If
this exceeds 0, then some debugging messages will be printed. This
value is passed down to related functions, but with 1 subtracted
for each pass.}
}
\value{
\code{\link[=dod.tideGauge]{dod.tideGauge()}} returns the full pathname of the
constructed file (in the CHS case) or the downloaded
file (in the NOAA case). This is a comma-separated file, with
first column named \code{"Time"}, and second column named
\code{"Height"}. FIXME: does NOAA put both in same file? If so,
we ought to name as such.
}
\description{
\code{\link[=dod.tideGauge]{dod.tideGauge()}} downloads tide-gauge data from either
Canadian or American tide gauges.
}
\details{
Downloads are done from for either the Canadian Hydrographic Service (CHS) or
from the American National Oceanographic and Atmospheric
Agency (NOAA), respectively (see References 1 and 2). The resultant data are saved in
either constructed filenames, or filenames provided by the user;
in either case, if a recent file already exists with the indicated name,
then no data are downloaded.

For NOAA files, water level and predictions are provided on the same
time sequence, but for CHS files, this is not the case, e.g. predictions
(in the author's tests) are on 15 minute intervals, starting at an
hour marker, but observations may be at a variety of times, depending
\code{start} and \code{end}.  Therefore, numerical comparision with CHS data
will require interpolation (see \sQuote{Examples}).
}
\examples{
\dontrun{
library(dod)
library(oce)
ofile <- dod.tideGauge(491)
pfile <- dod.tideGauge(491, "heightPredicted")
O <- read.csv(ofile)
O$time <- as.POSIXct(O$Date.Time, tz="UTC")
P <- read.csv(pfile)
P$time <- as.POSIXct(P$Date.Time, tz="UTC")
# Top panel: observation (black) and prediction (gray)
par(mfrow=c(2, 1))
oce.plot.ts(O$time, O$Water.Level, ylab="Water Level [m]")
lines(P$time, P$Predictions, col="gray", type="l")
# Bottom panel: misfit
Pinterp <- approxfun(P$time, P$Predictions)
oce.plot.ts(O$time, O$Water.Level-Pinterp(O$time), ylab="Deviation [m]")
}

}
\references{
\enumerate{
\item https://api-iwls.dfo-mpo.gc.ca/swagger-ui/index.html
\item https://api.tidesandcurrents.noaa.gov/api/prod/datagetter
}
}
