% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topo.R
\name{dod.topo}
\alias{dod.topo}
\title{Download topographic data}
\usage{
dod.topo(west = -66, east = -60, south = 43, north = 47, ...)
}
\arguments{
\item{west, east}{numeric values for the limits
of the data-selection box, in degrees. These
are converted to the -180 to 180 degree notation,
if needed. Then, west is rounded down to the nearest
1/100th degree, and east is rounded up to the the
nearest 1/100th degree. The results of these
operations are used in constructing the query for
the NOAA data server.}

\item{south, north}{latitude limits, treated in
a way that corresponds to the longitude limits.}

\item{...}{additional arguments passed to
\link[oce:download.topo]{oce::download.topo}.}
}
\value{
String indicating the full pathname to the downloaded file.
}
\description{
This function downloads topographic data using
\link[oce:download.topo]{oce::download.topo}.
}
