% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topo.R
\name{dod.topo}
\alias{dod.topo}
\title{Download topographic data}
\usage{
dod.topo(...)
}
\arguments{
\item{...}{arguments passed to
\code{\link[oce:download.topo]{oce::download.topo()}}.}
}
\value{
String indicating the full pathname to the downloaded file.
}
\description{
This function downloads topographic data using
\code{\link[oce:download.topo]{oce::download.topo()}}.
}
