% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug.R
\name{dodDebug}
\alias{dodDebug}
\title{Debug dod package functions}
\usage{
dodDebug(debug = 0, ...)
}
\arguments{
\item{debug}{integer value indicating level of debugging.
If this is less than 1, no debugging is done. Otherwise,
some functions will print debugging information.}

\item{...}{FILL IN not sure if this pulling from another
package or what.}
}
\description{
This function is used throughout the \code{dod} package to catch
bugs and provide assistance to users.
}
